avalon-backend/avalon/requirements.txt
aioredis==1.3.1
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
autopep8==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-redis==4.11.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
zope.interface==5.1.0
